# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DartShop API Documentation",
        "description": "API documentation for DartShop Laravel application",
        "contact": {
            "name": "DartShop Support",
            "email": "admin@dartshop.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/contact-messages": {
            "get": {
                "tags": [
                    "Admin/Contact Messages"
                ],
                "summary": "Get contact messages list (admin)",
                "description": "Retrieve all contact messages with admin filters and pagination",
                "operationId": "44450ce2fd4cf47758fe0cd732bf8313",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in name, email, subject, message",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (unread, read, replied)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "unread",
                                "read",
                                "replied"
                            ]
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "description": "Sort field (created_at, name, email, subject, status)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ContactMessage"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "title": "Pagination Meta",
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 12
                    },
                    "to": {
                        "type": "integer",
                        "example": 12
                    },
                    "total": {
                        "type": "integer",
                        "example": 60
                    },
                    "cache_hit": {
                        "type": "boolean",
                        "example": false
                    },
                    "cache_key": {
                        "type": "string",
                        "example": "products_list_abc123..."
                    },
                    "filters_available": {
                        "properties": {
                            "categories": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            },
                            "brands": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Brand"
                                }
                            },
                            "price_range": {
                                "properties": {
                                    "min": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "max": {
                                        "type": "number",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "Pagination Links",
                "description": "Pagination links",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://example.com/api/products?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://example.com/api/products?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://example.com/api/products?page=2"
                    }
                },
                "type": "object"
            },
...
            "ContactMessage": {
                "title": "Contact Message",
                "description": "Contact form message",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Question about products"
                    },
                    "message": {
                        "type": "string",
                        "example": "I have a question about..."
                    },
                    "status": {
                        "type": "string",
                        "example": "unread"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Brand": {
                "title": "Brand",
                "description": "Product brand model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Harrows"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Product category model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dart Flights"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}